+++ NEWS.md
@@ -1,12 +1,16 @@
 # wbdataset
 
+# 0.1.1035
+* `get_property_definition()` works with CSRF, `check_existing_property()` returns
+correct datatype. Improved documentation and unit tests.
+
 # 0.1.1034
 * Remove the unnecessary `log_path` parameter from the create and copy function 
 families.
 * Many corrections in the formatting and the contents of the function reference.
  
 # wbdataset 0.1.1023
-* [check_existing_item()], [check_existing_property()]
+* `check_existing_item()`, `check_existing_property()`
 
 # wbdataset 0.1.1014
 * copy_wikidata_item, copy_wikidata_property works with one or several items/properties